fix: prevent segfaults from race conditions and integer overflow
1. Worker.cpp: Fix TOCTOU race condition in postMessage()
   - Check both TerminatedFlag AND TerminateRequestedFlag before posting
   - Previously only checked TerminatedFlag, allowing messages to be posted
     to a worker that was in the process of terminating
   - This caused use-after-free when the task executed on a dead context

2. JSSQLStatement.cpp: Fix integer overflow in sqlite3_serialize()
   - Changed cast from unsigned int to size_t for buffer length
   - sqlite3_int64 (64-bit) was being truncated to unsigned int (32-bit)
   - Databases >2GB would segfault due to buffer size mismatch

Fixes: oven-sh/bun#25610 (Worker thread race condition)
Fixes: oven-sh/bun#19470 (SQLite segfault on large datasets)

+++ src/bun.js/bindings/sqlite/JSSQLStatement.cpp
@@ -1280,7 +1280,7 @@ JSC_DEFINE_HOST_FUNCTION(jsSQLStatementSerialize, (JSC::JSGlobalObject * lexical
         return {};
     }
 
-    RELEASE_AND_RETURN(scope, JSBuffer__bufferFromPointerAndLengthAndDeinit(lexicalGlobalObject, reinterpret_cast<char*>(data), static_cast<unsigned int>(length), NULL, sqlite_free_typed_array));
+    RELEASE_AND_RETURN(scope, JSBuffer__bufferFromPointerAndLengthAndDeinit(lexicalGlobalObject, reinterpret_cast<char*>(data), static_cast<size_t>(length), NULL, sqlite_free_typed_array));
 }
 
 JSC_DEFINE_HOST_FUNCTION(jsSQLStatementLoadExtensionFunction, (JSC::JSGlobalObject * lexicalGlobalObject, JSC::CallFrame* callFrame))
